Adding insertion method

+++ app/VirtualStore/Entities/Product.cs
@@ -2,14 +2,14 @@
 {
     public class Product
     {
-        public int ID { get; private set; }
+        public long ID { get; private set; }
         public string Name { get; set; }
         public Category Category { get; set; }
         public Manufacturer Manufacturer { get; set; }
         public decimal Price { get; set; }
         public string Description { get; set; }
         public Product() { }
-        public Product(int iD, string name, Category category, Manufacturer manufacturer, decimal price, string description)
+        public Product(long iD, string name, Category category, Manufacturer manufacturer, decimal price, string description)
         {
             this.ID = iD;
             this.Name = name;
